Move source files to src/ directory
